Break up and Refactor driver
