+++ Makefile
@@ -42,7 +42,6 @@ CFLAGS += -DVERSION=\"v0.9.0.0-$(GIT_VERSION)\"
 MAINLOGIC_SOURCE := \
 	main.c \
 	device.c \
-	driver.c \
 	print.c \
 	logic/options.c \
 	logic/options_fan.c \
@@ -54,6 +53,16 @@ MAINLOGIC_SOURCE := \
 	logic/settings/hydro_asetekpro.c \
 	logic/settings/hydro_coolit.c \
 	logic/settings/psu.c
+
+DRIVER_SOURCE := \
+	driver/asetek.c \
+	driver/asetekpro.c \
+	driver/commanderpro.c \
+	driver/coolit_old.c \
+	driver/coolit.c \
+	driver/dongle.c \
+	driver/rmi.c \
+	driver/unsupported.c
 
 LOWLEVEL_SOURCE := \
 	lowlevel/asetek.c \
@@ -95,6 +104,9 @@ EXECUTABLE := OpenCorsairLink.elf
 OBJS := \
 	${MAINLOGIC_SOURCE:.c=.o}
 
+OBJS_DRV := \
+	${DRIVER_SOURCE:.c=.o}
+
 OBJS_LL := \
 	${LOWLEVEL_SOURCE:.c=.o} 
 
@@ -111,16 +123,16 @@ default: all
 rebuild: clean all
 
 .PHONY: style
-style: $(MAINLOGIC_SOURCE) $(LOWLEVEL_SOURCE) $(PROTOCOL_SOURCE) $(HEADER)
+style: $(MAINLOGIC_SOURCE) $(DRIVER_SOURCE) $(LOWLEVEL_SOURCE) $(PROTOCOL_SOURCE) $(HEADER)
 	clang-format -style=file -i $^
 
 .PHONY: tidy
-tidy: $(MAINLOGIC_SOURCE) $(LOWLEVEL_SOURCE) $(PROTOCOL_SOURCE)
+tidy: $(MAINLOGIC_SOURCE) $(DRIVER_SOURCE) $(LOWLEVEL_SOURCE) $(PROTOCOL_SOURCE)
 	clang-tidy $^ -checks=* -- -std=c++11
 
 .PHONY: clean
 clean:
-	$(RM) $(EXECUTABLE) $(OBJS) $(OBJS_LL) $(OBJS_PROTO)
+	$(RM) $(EXECUTABLE) $(OBJS) $(OBJS_DRV) $(OBJS_LL) $(OBJS_PROTO)
 
 .PHONY: install
 install: $(EXECUTABLE)
@@ -134,7 +146,7 @@ uninstall:
 %.o: %.c
 	$(CC) $(CFLAGS) -g -c -o $@ $<
 
-$(EXECUTABLE): $(OBJS) $(OBJS_PROTO) $(OBJS_LL)
+$(EXECUTABLE): $(OBJS) $(OBJS_DRV) $(OBJS_PROTO) $(OBJS_LL)
 	$(CC) $^ $(CFLAGS) $(LDFLAGS) -o $@
 
 ####################################################################################################
