+++ driver/unsupported.c
@@ -1,0 +1,47 @@
+/*
+ * This file is part of OpenCorsairLink.
+ * Copyright (C) 2017-2019  Sean Nelson <audiohacked@gmail.com>
+
+ * OpenCorsairLink is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 2 of the License, or
+ * any later version.
+
+ * OpenCorsairLink is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+
+ * You should have received a copy of the GNU General Public License
+ * along with OpenCorsairLink.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "driver.h"
+#include "device.h"
+#include "print.h"
+
+int
+corsairlink_unsupported_led(
+    struct corsair_device_info* dev, struct libusb_device_handle* handle, struct led_control* ctrl )
+{
+    msg_info( "Unsupported Generic LED Function\n" );
+    return 0;
+}
+
+int
+corsairlink_unspported_fan(
+    struct corsair_device_info* dev, struct libusb_device_handle* handle, struct fan_control* ctrl )
+{
+    msg_info( "Unsupported Generic Fan Function\n" );
+    return 0;
+}
+
+int
+corsairlink_unspported_pump(
+    struct corsair_device_info* dev,
+    struct libusb_device_handle* handle,
+    struct pump_control* ctrl )
+{
+    msg_info( "Unsupported Generic Pump Function\n" );
+    return 0;
+}
